+++ lib/online_migrations/error_messages.rb
@@ -149,17 +149,42 @@ It will use a combination of a VIEW and column aliasing to work with both column
   end
 
 4. Replace usages of the old column with a new column in the codebase
+<% if enumerate_columns_in_select_statements %>
+5. Ignore old column
+
+<% if ar_version >= 5 %>
+  self.ignored_columns = [:<%= column_name %>]
+<% else %>
+  def self.columns
+    super.reject { |c| c.name == \"<%= column_name %>\" }
+  end
+<% end %>
+
+6. Deploy
+7. Remove the column rename config from step 1
+8. Remove the column ignore from step 5
+9. Remove the VIEW created in step 3 and finally rename the column:
+
+  class Finalize<%= migration_name %> < <%= migration_parent %>
+    def change
+      finalize_column_rename :<%= table_name %>, :<%= column_name %>, :<%= new_column %>
+    end
+  end
+
+10. Deploy
+<% else %>
 5. Deploy
 6. Remove the column rename config from step 1
 7. Remove the VIEW created in step 3 and finally rename the column:
 
   class Finalize<%= migration_name %> < <%= migration_parent %>
     def change
-      finalize_column_rename <%= table_name.inspect %>, <%= column_name.inspect %>, <%= new_column.inspect %>
+      finalize_column_rename :<%= table_name %>, :<%= column_name %>, :<%= new_column %>
     end
   end
 
-8. Deploy",
+8. Deploy
+<% end %>",
 
       change_column_with_not_null:
 "Changing the type is safe, but setting NOT NULL is not.",